Platform, ZCU10X: Account for ATF in memory layout.
The Arm Trusted Firmware shipped in the default Xilinx BOOT.BIN is
loaded in a non secure zone, more specifically at the end of the first
memory bank. Being in a non-secure zone means that can be easily
overwritten.

The memory size is therefore shrunk to take this into account and avoid
memory corruption.