mem: add coherency request to address spaces.
Memory coherency is not always provided by hardware. For example arm
SMMU, on some platform, do not have coherent page table walks, and that
makes sharing page tables between MMU and SMMU have "undefined
behaviour".

This commit implements a way to force coherency in an address space
when it requires so.