fix race condition on psci state
It was possible for a cpu to see its vcpu psci in a transitory state
when returning from the hypervisor. This might happen when while
performing some operation with its vcpu off, another cpu emits a cpu on
psci message. Theres a race condition since the moment the cpu has
finished processing messages might happen after the moment the other cpu
sets the vcpu state to "on pending" which results in the cpu arriving at
the vcpu_run operation with the vcpu in this transitory psci state. By
removing the check for the off state, the cpu will try to enter idle
even if the state is transitory, and wake up imediately since an ipi
will still be pending, forcing a re-check of the messages which will result
in bringing the vcpu to the on state.