arm/cache: fix flush operation
The flush operation was only cleaning caches and not invalidating them
which became a problem in some cores when waking up from powerdown and
running from physical memory. When updating a pte directly in physical
memory, a stale entry might still be in cache resulting in a page-table
walk abort when enabling virtual memory.