armv8: allow arbitrary master (boot) cpu
Previously we assumed all arm platforms first cpu to boot would be cpu 0
with mpidr[15:0] == 0 which would later become master cpu and be
responsible for most initializations. This is not true for some platforms
(eg: nvidia tx2). We now allow any of the cpus to be this cpu, but still
assume it boots alone and remaining cpus are only woke up after it emits
psci cpu_on requests for them.

Co-authored-by: Gero Schwäericke <gero.schwaericke@tum.de>