allow cache topology to be in platform description
Previsouly in arm platforms we used system registers to discover the
cache topology. Because riscv does not have this capability and has to
rely on device trees for which we do not have a parser and dont intend
it, a platform port of riscv should add the cache topology description
to the platform descprition.