copy vm image to a new memory region by default
As an alternative as using the vm image inplace (i.e. where it was
loaded). Although this might slow down VM boot, it allows to use
superpages for the complete VM image which improves performance. It will
also allow reuse of the image in features features such as VM reboot.