fix(mem/mmu): preserve sw reserved flags when expanding pt entry
The only flag in use currently is the "reserved" flag, which means
this entry was previously reserved for mapping. If the entry is
expanded to a next level table and its entries are not marked also
as such, they can be later re-allocated by another thread/cpu resulting
in an mapping overlap.

Signed-off-by: Jose Martins <josemartins90@gmail.com>