fix(hypercall/risc-v): don't exclude A0 register from args retrieval in hypercall()
The current multi-arch implementation does not include RISC-V A0 register
when running Bao on top of a RISC-V arch. A0 contains the ipc_id and it is
not passed to the ipc_ipercall() function causing possible errors due to the
right argument missing.

To fix this, the HYPCALL_ARG_REG(ARG) now starts with offset 0 (now A0 is the first
register considered for RISC-V), but then the same interface function must be changed
for ARM (by adding +1 to the updated offset).

To improve readability and avoid misunderstandings, the argument retrieved by A0 in
RISC-V and by X1 in ARM is now called ipc_id.

Signed-off-by: Filippo Fontana <filippofontana1998@gmail.com>