feat(ci): include ci.mk in main Makefile
When including ci.mk in the Makefile it is needed to check if the target
rules are rules that will trigger a build or not as most CI rules don't
require a build. Otherwise many operations needded only for a build will
be triggered even if we are only, for example, checking code formatting.

Signed-off-by: Jose Martins <josemartins90@gmail.com>