fix(empty-structs): Add dummy byte to empty structs
Add a dummy uint8_t to the structs page_table_arch and iommu_vm_arch since
C11 does not allow structs to be declared without members.
This was the best solution available avoiding majors modifications to bao.
Considering both structs are only used in a small number of ocasions, the
increase in memory usage is acceptable.

Signed-off-by: Miguel Silva <miguelafsilva5@gmail.com>