fix(cpu_idle): replace psci_power_down by a WFI call
For physical boards (e.g. ZCU102), when using the Remote IO
there is a problem already identified when the Frontend VM's
CPU goes to the idle state, as a result of the trap process
and MMIO emulation. In this scenario, currently Bao Hypervisor
is unable to save the state of the registers before powering off.
This patch replaces the psci_power_down call by an explicit WFI
(Wait For Interrupt), temporarily resolving the problem.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>