fix(core/remote_io): frontend VM with multiple vCPUs
For some heavy workloads (e.g. iperf3) when a frontend VM
is configured with more than one vCPU, two Remote I/O requests
can be raised practically at the same time, causing the
Dispatcher I/O system (implemented in the backend VM Linux kernel)
to miss a Remote I/O interrupt. In this case, as we are unable to
change the way the Linux kernel handles interrupts, the solution is
to return the number of requests that are still in the pending phase
at the time of the Ask hypercall, so that the Dispatcher I/O system
can do the remaining `n` Ask hypercalls to process the remaining orders.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>