feat: finish interrupts on cpu idle
This allows a cpu to enter idle state after hanlding an interrupt, while
leaving the interrupt in the correct state. If this was not carried out,
the cpu might not wake up on another interrupt because it is still
handling the current interrupt or the same interrupt might not be
triggered again.

Signed-off-by: Jose Martins <josemartins90@gmail.com>