update(ci): allow for submodule not be cloned
If this reposiroty is not cloned with '--recursive', the ci subdmoudle
will not be prenset which will result in any invocation of make to fail.
This modification allows that if the ci submodule is not present, it
will simply be ignored and typical make calls for building the
hypervisor will continue normally.

Signed-off-by: Jose Martins <josemartins90@gmail.com>