fix(armv8/cpu): replace psci_power_down by a WFI call
For ARMv8-A physical boards (e.g. ZCU102), when using the Remote I/O
there is a problem already identified when the Frontend VM's
CPU goes to the idle state, as a result of the trap process
and MMIO emulation. In this scenario, currently Bao Hypervisor
is unable to save the state of the registers before powering off.
This commit replaces the psci_power_down call by an explicit WFI
(Wait For Interrupt), similar to ARMv8-R and RISC-V, temporarily
resolving the problem.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>