feat(armv8/psci): Handle unexpected powerdown behavior
We've observed that some platforms behave unexpectedly when performing
power down. In these cases, after powerdown, the CPU cores are not awaken
by interrupts as expected by Bao. This commit solve this, by adding the
ability to skip the powerdown firmware call, instead relying on the
fallback mechanism of using standby.

Signed-off-by: David Cerdeira <davidmcerdeira@gmail.com>