fix(linker): Extra memory removed from dmem_phys_beg
For MMU systems we manage the location counter to relocate VM images virtual
address space.
This allows us to maintain a contiguous virtual address for each cpu between
the hypervisor image and it’s private memory. However, for non-mmu systems
this operation results in extra memory being erroneously allocated for the vms,
as this space is already allocated between vm_image_start and vm_image_end.

Signed-off-by: Miguel Silva <miguelafsilva5@gmail.com>