fix(riscv/interrupts): add missing initialization for irqc_timer_int_id
irqc_timer_int_id is the variable representing the ID of the timer
interrupt. It is declared in the interrupts.c but is never defined,
causing a bug while handling timer interrupts inside the handler for
the AIA controller ("no interrupt handler for ID 0").

The irqc_timer_int_id variable is initialized following the same
strategy used for interrupts_ipi_id.

Signed-off-by: Filippo Fontana <filippo.fontana@minervasys.tech>