fix(riscv/irqc): align IMSIC handler to AIA specifications
"The RISC-V Advanced Interrupt Architecture" specifications explain
in Chapter 3 (pages 26-27) how interrupts should be handled with IMSIC,
in particular, interrupt claiming has to be done right after reading the
STOPEI register or in the same read instruction with the csrrw
instruction.

This commit aligns the IMSIC interrupt handler to that specs.
Without this change, if an interrupt with higher priority becomes
pending while handling a lower one, writing to STOPEI after
`interrupts_handle()` executes will claim the higher IRQ, freezing
the execution.

Signed-off-by: Filippo Fontana <filippo.fontana@minervasys.tech>