fix(armv8/aarch32): restore/save extra lr on vm entry/exit
Hyp mod uses lr from user mode. This means we were only saving/restoring
the lr from this mode. But in addition to this lr we need to restore the
lr of the mode we just preemted (e.g., lr_irq) so that we can inspect
and manipulate the state of the mode that was trapped.

Signed-off-by: Jose Martins <josemartins90@gmail.com>