fix(armv8/aarch32): correctly access state of different el1 modes
We need to take into account the fact that hyp uses el1's lr_usr but
when accessing the guest's current cuntext which lr we access/update
might be on a different el1 mode.

Plus, we were ignoring the FIQ mode. Although the guest will not receive
FIQs because the hypervisor prevents it, it can still use the FIQ
context.

Signed-off-by: Jose Martins <josemartins90@gmail.com>