ref(mem): change bitmaps to static allocation
Previously we allocated bitmaps in run-time which lead to waste
of MPU regions in MMU-less platforms. This commit creates a single
array bitmap with the size of the declared memory platform regions.
This array is allocated in .bss section and, therefore, is reserved
as part of the hyp image (pp_root_reserve_hyp_mem).

Signed-off-by: Daniel Oliveira <drawnpoetry@gmail.com>