fix: vm lock and sync initialization
The vm struct initizalization initializes its internal synchronization
primitives, however, itself depends on this primitives. This commit
creates a sync primitive during vm assignment and passes it to the
vm_init function so it can use it for synchronization before the
internal vm sync primitives are initialized.

Also, it fixes the initialization of the lock used for vm cpu
assignment.

Signed-off-by: Jose Martins <josemartins90@gmail.com>