feat(make): make sure executable is not position independent
Bao should be compile as position dependent code. Originally we have the
-fno-pic option but this prevents only position independent for library
binaties (e.g., .so). We want to make sure the final executable is not position
independent (albeit this is the default option, we should make it explicit).
For this we need to add the no-pie options to compiler and linker flags.

Signed-off-by: Jose Martins <josemartins90@gmail.com>