ref(vm_mem_init): move vm_mem_prot_init to later in vm_init
The slave side mmio protection requires the master VM to map the mmio regions
in all vcpus. This is done in the as_init function, as it works for the hyp
as and vm as. However, before, it was done in vm_master_init and at this point
the vm and vcpu structs were not correctly filled, making the broadcast of mmio
mpu regions incorrect. The solution in this commit, simply moves the
vm_mem_prot to right after the vcpu init, where everything is already
initialized.

Signed-off-by: Miguel Silva <miguelafsilva5@gmail.com>